requirements.txt
# ==================================================
# Dependências do Portal PGE-MS
# ==================================================

# Framework Web
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
werkzeug>=3.0.0

# Templates
Jinja2>=3.1.0

# Banco de Dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0  # Migrations de banco de dados

# Autenticação
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0

# Variáveis de ambiente
python-dotenv>=1.0.1

# HTTP Requests (para APIs externas)
requests>=2.32.0
httpx[http2]>=0.27.0
aiohttp>=3.9.0
nest_asyncio>=1.6.0

# Rate Limiting
slowapi>=0.1.9  # SECURITY: Proteção contra DoS/brute-force

# Logging Estruturado
structlog>=24.1.0  # Logs estruturados com JSON, contexto e request_id

# Parser XML (TJ-MS)
lxml>=5.2.0
defusedxml>=0.7.1  # SECURITY: Proteção contra XXE

# Geração de documentos
python-docx>=1.1.0
docx2pdf>=0.1.8  # Conversão DOCX->PDF via MS Word (Windows)

# PDF e imagens (Matrículas Confrontantes e Agente TJ-MS)
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.10
Pillow>=10.0.0

# Busca vetorial (embeddings)
numpy>=1.24.0

# BERT Training (análise de dados)
pandas>=2.0.0
openpyxl>=3.1.0

# Scrapper de Subconta (Prestação de Contas)
playwright>=1.40.0
rich>=13.0.0

# ==================================================
# Instalação
# ==================================================
# pip install -r requirements.txt
# playwright install chromium  # Após instalar playwright
pytz>=2024.1
